Edit introduction
I have edited the introduction to directly address Reviewer 2's concern about the lack of a "particularly compelling case." I have removed the "refresh" language. But more specifically, I have highlighted that this is not just a data refresh but also that the new modeling approach and the new data allow me to address several shortcomings of prior work. I also include an explicit paragraph about the birthplace and language endogamy issue to help bring that to the fore.