Edit birthplace and language endogamy section
I have revised this section considerably to incorporate the actual analysis of birthplace and language endogamy that used to be in the results section. Although its a little weird to incorporate it before explaining the data sources, I think it does a much better job of literally illustrating the issue when it is done here.